separate bridge and bridge integration tests, move deployBridge
